add math steps for solving in app
