+++ app/src/main/java/org/stepic/droid/view/fragments/StringStepFragment.java
@@ -1,19 +1,7 @@
 package org.stepic.droid.view.fragments;
 
-import android.content.Context;
-import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.view.KeyEvent;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.view.inputmethod.EditorInfo;
-import android.widget.EditText;
-import android.widget.TextView;
-
 import com.squareup.otto.Subscribe;
 
-import org.stepic.droid.R;
 import org.stepic.droid.events.InternetIsEnabledEvent;
 import org.stepic.droid.events.attempts.FailAttemptEvent;
 import org.stepic.droid.events.attempts.SuccessAttemptEvent;
@@ -21,49 +9,14 @@ import org.stepic.droid.events.submissions.FailGettingLastSubmissionEvent;
 import org.stepic.droid.events.submissions.FailSubmissionCreatedEvent;
 import org.stepic.droid.events.submissions.SubmissionCreatedEvent;
 import org.stepic.droid.events.submissions.SuccessGettingLastSubmissionEvent;
-import org.stepic.droid.model.Attempt;
 import org.stepic.droid.model.Reply;
 
-public class StringStepFragment extends StepWithAttemptsFragment {
-
-    EditText mAnswerField;
-
-    @Nullable
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        View v = super.onCreateView(inflater, container, savedInstanceState);
-        mAnswerField = (EditText) ((LayoutInflater) this.getActivity().getSystemService(Context.LAYOUT_INFLATER_SERVICE)).inflate(R.layout.view_string_attempt, mAttemptContainer, false);
-        mAttemptContainer.addView(mAnswerField);
-        mAnswerField.setOnEditorActionListener(new TextView.OnEditorActionListener() {
-            @Override
-            public boolean onEditorAction(TextView v, int actionId, KeyEvent event) {
-                boolean handled = false;
-                if (actionId == EditorInfo.IME_ACTION_SEND) {
-                    mActionButton.performClick();
-                    handled = true;
-                }
-                return handled;
-            }
-        });
-        return v;
-    }
-
-
-    @Override
-    protected void showAttempt(Attempt attempt) {
-        mAnswerField.getText().clear();
-    }
-
+public class StringStepFragment extends OneLineWithSendStepFragment {
     @Override
     protected Reply generateReply() {
         return new Reply.Builder()
                 .setText(mAnswerField.getText().toString())
                 .build();
-    }
-
-    @Override
-    protected void blockUIBeforeSubmit(boolean needBlock) {
-        mAnswerField.setEnabled(!needBlock);
     }
 
     @Override
